Container/pythonACI/requirements.txt
azure-storage-blob==1.4.0
geopy==1.18.1
shapely==1.6.4.post2
pyproj==1.9.5.1
pandas==0.23.4